pyproject.toml
[project]
name = "caal"
version = "0.1.0"
description = "CAAL - Voice assistant with n8n workflow integrations"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "CoreWorxLab", email = "cmac@coreworxlab.com"}
]
requires-python = ">=3.10"
keywords = [
    "livekit",
    "voice",
    "speech-to-text",
    "text-to-speech",
    "mcp",
    "n8n",
    "ollama",
    "speaches",
    "local",
    "self-hosted"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # LiveKit core
    "livekit-agents[silero,turn-detector,mcp,groq]~=1.3",
    # LLM
    "ollama>=0.3.0",
    "groq>=0.11.0",  # For GroqProvider (livekit-plugins-groq doesn't include it)
    "anthropic>=0.40.0",  # For AnthropicProvider
    # Core utilities
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",  # For model preloading
    # Webhook server
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    # Web search
    "ddgs>=9.0.0",
    # Wake word detection
    "openwakeword>=0.6.0",
]

[project.optional-dependencies]
cloud = [
    "livekit-plugins-deepgram~=1.0",
    "livekit-plugins-cartesia~=1.0",
    "livekit-plugins-openai~=1.0",
]

[project.urls]
Homepage = "https://github.com/CoreWorxLab/caal"
Repository = "https://github.com/CoreWorxLab/caal"
Issues = "https://github.com/CoreWorxLab/caal/issues"
Documentation = "https://github.com/CoreWorxLab/caal#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/caal"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
